+++ packages/@uppy/core/src/Uppy.ts
@@ -13,7 +13,17 @@ import DefaultStore from '@uppy/store-default'
 import getFileType from '@uppy/utils/lib/getFileType'
 import getFileNameAndExtension from '@uppy/utils/lib/getFileNameAndExtension'
 import { getSafeFileId } from '@uppy/utils/lib/generateFileID'
-import type { UppyFile, Meta, Body } from '@uppy/utils/lib/UppyFile'
+import type {
+  UppyFile,
+  Meta,
+  Body,
+  MinimalRequiredUppyFile,
+} from '@uppy/utils/lib/UppyFile'
+import type { CompanionFile } from '@uppy/utils/lib/CompanionFile'
+import type {
+  CompanionClientProvider,
+  CompanionClientSearchProvider,
+} from '@uppy/utils/lib/CompanionClientProvider'
 import type {
   FileProgressNotStarted,
   FileProgressStarted,
@@ -37,8 +47,7 @@ import packageJson from '../package.json'
 import locale from './locale.ts'
 
 import type BasePlugin from './BasePlugin.ts'
-import type UIPlugin from './UIPlugin.ts'
-import type { Restrictions } from './Restricter.ts'
+import type { Restrictions, ValidateableFile } from './Restricter.ts'
 
 type Processor = (fileIDs: string[], uploadID: string) => Promise<void> | void
 
@@ -46,29 +55,82 @@ type FileRemoveReason = 'user' | 'cancel-all'
 
 type LogLevel = 'info' | 'warning' | 'error' | 'success'
 
-export type UnknownPlugin<M extends Meta, B extends Body> = InstanceType<
-  typeof BasePlugin<any, M, B> | typeof UIPlugin<any, M, B>
->
+export type UnknownPlugin<
+  M extends Meta,
+  B extends Body,
+  PluginState extends Record<string, unknown> = Record<string, unknown>,
+> = BasePlugin<any, M, B, PluginState>
 
-type UnknownProviderPlugin<M extends Meta, B extends Body> = UnknownPlugin<
-  M,
-  B
-> & {
-  provider: {
-    logout: () => void
+export type UnknownProviderPluginState = {
+  authenticated: boolean | undefined
+  breadcrumbs: {
+    requestPath: string
+    name: string
+    id?: string
+  }[]
+  didFirstRender: boolean
+  currentSelection: CompanionFile[]
+  filterInput: string
+  loading: boolean | string
+  folders: CompanionFile[]
+  files: CompanionFile[]
+  isSearchVisible: boolean
+}
+/*
+ * UnknownProviderPlugin can be any Companion plugin (such as Google Drive).
+ * As the plugins are passed around throughout Uppy we need a generic type for this.
+ * It may seems like duplication, but this type safe. Changing the type of `storage`
+ * will error in the `Provider` class of @uppy/companion-client and vice versa.
+ *
+ * Note that this is the *plugin* class, not a version of the `Provider` class.
+ * `Provider` does operate on Companion plugins with `uppy.getPlugin()`.
+ */
+export type UnknownProviderPlugin<
+  M extends Meta,
+  B extends Body,
+> = UnknownPlugin<M, B, UnknownProviderPluginState> & {
+  onFirstRender: () => void
+  title: string
+  files: UppyFile<M, B>[]
+  icon: () => JSX.Element
+  provider: CompanionClientProvider
+  storage: {
+    getItem: (key: string) => Promise<string | null>
+    setItem: (key: string, value: string) => Promise<void>
+    removeItem: (key: string) => Promise<void>
   }
 }
 
-// The user facing type for UppyFile used in uppy.addFile() and uppy.setOptions()
-export type MinimalRequiredUppyFile<M extends Meta, B extends Body> = Required<
-  Pick<UppyFile<M, B>, 'name' | 'data' | 'type' | 'source'>
-> &
-  Partial<
-    Omit<UppyFile<M, B>, 'name' | 'data' | 'type' | 'source' | 'meta'>
-    // We want to omit the 'meta' from UppyFile because of internal metadata
-    // (see InternalMetadata in `UppyFile.ts`), as when adding a new file
-    // that is not required.
-  > & { meta?: M }
+/*
+ * UnknownSearchProviderPlugin can be any search Companion plugin (such as Unsplash).
+ * As the plugins are passed around throughout Uppy we need a generic type for this.
+ * It may seems like duplication, but this type safe. Changing the type of `title`
+ * will error in the `SearchProvider` class of @uppy/companion-client and vice versa.
+ *
+ * Note that this is the *plugin* class, not a version of the `SearchProvider` class.
+ * `SearchProvider` does operate on Companion plugins with `uppy.getPlugin()`.
+ */
+export type UnknownSearchProviderPluginState = {
+  isInputMode?: boolean
+  searchTerm?: string | null
+} & Pick<
+  UnknownProviderPluginState,
+  | 'loading'
+  | 'files'
+  | 'folders'
+  | 'currentSelection'
+  | 'filterInput'
+  | 'didFirstRender'
+>
+export type UnknownSearchProviderPlugin<
+  M extends Meta,
+  B extends Body,
+> = UnknownPlugin<M, B, UnknownSearchProviderPluginState> & {
+  onFirstRender: () => void
+  title: string
+  icon: () => JSX.Element
+  provider: CompanionClientSearchProvider
+}
 
 interface UploadResult<M extends Meta, B extends Body> {
   successful?: UppyFile<M, B>[]
@@ -775,8 +837,8 @@ export class Uppy<M extends Meta, B extends Body> {
   }
 
   validateRestrictions(
-    file: UppyFile<M, B>,
-    files = this.getFiles(),
+    file: ValidateableFile<M, B>,
+    files: ValidateableFile<M, B>[] = this.getFiles(),
   ): RestrictionError<M, B> | null {
     try {
       this.#restricter.validate(files, [file])
@@ -813,9 +875,12 @@ export class Uppy<M extends Meta, B extends Body> {
     const { allowNewUpload } = this.getState()
 
     if (allowNewUpload === false) {
-      const error = new RestrictionError(this.i18n('noMoreFilesAllowed'), {
-        file,
-      })
+      const error = new RestrictionError<M, B>(
+        this.i18n('noMoreFilesAllowed'),
+        {
+          file,
+        },
+      )
       this.#informAndEmit([error])
       throw error
     }
@@ -1841,7 +1906,7 @@ export class Uppy<M extends Meta, B extends Body> {
    * Passes messages to a function, provided in `opts.logger`.
    * If `opts.logger: Uppy.debugLogger` or `opts.debug: true`, logs to the browser console.
    */
-  log(message: string | Record<string, unknown> | Error, type?: string): void {
+  log(message: string | Record<any, any> | Error, type?: string): void {
     const { logger } = this.opts
     switch (type) {
       case 'error':